Fixed issue 385: Move OptimizeRunOrderSuite from spock-core to spock-maven to solve a problem with Android's test runner
